+++ src/librawspeed/decoders/DngDecoder.cpp
@@ -810,7 +810,7 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
 
     for (int y = 0; y < 2; y++) {
       for (int x = 0; x < 2; x++)
-        mRaw->blackLevelSeparate[y * 2 + x] = value;
+        mRaw->blackLevelSeparate[y * 2 + x] = implicit_cast<int>(value);
     }
   } else {
     for (int y = 0; y < 2; y++) {
@@ -822,7 +822,7 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
             static_cast<double>(value) > std::numeric_limits<BlackType>::max())
           ThrowRDE("Error decoding black level");
 
-        mRaw->blackLevelSeparate[y * 2 + x] = value;
+        mRaw->blackLevelSeparate[y * 2 + x] = implicit_cast<int>(value);
       }
     }
   }
@@ -844,7 +844,8 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
           static_cast<double>(value) > std::numeric_limits<BlackType>::max())
         ThrowRDE("Error decoding black level");
 
-      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i], value,
+      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i],
+                                  implicit_cast<int>(value),
                                   &mRaw->blackLevelSeparate[i]))
         ThrowRDE("Integer overflow when calculating black level");
     }
@@ -866,7 +867,8 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
           static_cast<double>(value) > std::numeric_limits<BlackType>::max())
         ThrowRDE("Error decoding black level");
 
-      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i], value,
+      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i],
+                                  implicit_cast<int>(value),
                                   &mRaw->blackLevelSeparate[i]))
         ThrowRDE("Integer overflow when calculating black level");
     }
